Move : move ReadOnly&Search Autocomplete code #19
